requirements.txt
tkinter
Pillow
pyperclip
ctypes
pickle